project_1/rag_solutions/requirements.txt
sentence-transformers
faiss-cpu
groq
streamlit
rank-bm25
numpy
matplotlib
pandas
